drivers: cc110x: fix interrupt logic
