+++ Makefile.dep
@@ -558,5 +558,6 @@ ifneq (,$(filter random,$(USEMODULE)))
 endif
 
 ifneq (,$(filter cc110x,$(USEMODULE)))
+    USEMODULE += xtimer
     DISABLE_MODULE += gnrc_sixlowpan_iphc
 endif
